Creating a session for the user.  Authenticating google id tokens
When the user successfully logs into google, send their ID token to the
server.  The server verifies this token is not forged or manipulated,
then creates a session containing their unique google user ID.  From
then on, the user is allowed to navigate to authenticated URLs.  Only
the base URL and /api/auth are not authenticated.